hzn exchange service command extract namespace attribute from clusterDeploymentString
