+++ kube_operator/client.go
@@ -93,7 +93,7 @@ func NewDynamicKubeClient() (dynamic.Interface, error) {
 
 // Install creates the objects specified in the operator deployment in the cluster and creates the custom resource to start the operator
 func (c KubeClient) Install(tar string, envVars map[string]string, agId string, crInstallTimeout int64) error {
-	apiObjMap, namespace, err := processDeployment(tar, envVars, agId, crInstallTimeout)
+	apiObjMap, namespace, err := ProcessDeployment(tar, envVars, agId, crInstallTimeout)
 	if err != nil {
 		return err
 	}
@@ -151,7 +151,7 @@ func (c KubeClient) Install(tar string, envVars map[string]string, agId string, 
 
 // Install creates the objects specified in the operator deployment in the cluster and creates the custom resource to start the operator
 func (c KubeClient) Uninstall(tar string, agId string) error {
-	apiObjMap, namespace, err := processDeployment(tar, map[string]string{}, agId, 0)
+	apiObjMap, namespace, err := ProcessDeployment(tar, map[string]string{}, agId, 0)
 	if err != nil {
 		return err
 	}
@@ -184,7 +184,7 @@ func (c KubeClient) Uninstall(tar string, agId string) error {
 	return nil
 }
 func (c KubeClient) OperatorStatus(tar string, agId string) (interface{}, error) {
-	apiObjMap, namespace, err := processDeployment(tar, map[string]string{}, agId, 0)
+	apiObjMap, namespace, err := ProcessDeployment(tar, map[string]string{}, agId, 0)
 	if err != nil {
 		return nil, err
 	}
@@ -200,7 +200,7 @@ func (c KubeClient) OperatorStatus(tar string, agId string) (interface{}, error)
 	return status, nil
 }
 func (c KubeClient) Status(tar string, agId string) ([]ContainerStatus, error) {
-	apiObjMap, namespace, err := processDeployment(tar, map[string]string{}, agId, 0)
+	apiObjMap, namespace, err := ProcessDeployment(tar, map[string]string{}, agId, 0)
 	if err != nil {
 		return nil, err
 	}
@@ -244,8 +244,8 @@ func (c KubeClient) Status(tar string, agId string) ([]ContainerStatus, error) {
 	}
 }
 
-// processDeployment takes the deployment string and converts it to a map with the k8s objects, the namespace to be used, and an error if one occurs
-func processDeployment(tar string, envVars map[string]string, agId string, crInstallTimeout int64) (map[string][]APIObjectInterface, string, error) {
+// ProcessDeployment takes the deployment string and converts it to a map with the k8s objects, the namespace to be used, and an error if one occurs
+func ProcessDeployment(tar string, envVars map[string]string, agId string, crInstallTimeout int64) (map[string][]APIObjectInterface, string, error) {
 	// Read the yaml files from the commpressed tar files
 	yamls, err := getYamlFromTarGz(tar)
 	if err != nil {
